Implement get_assets_def (#30377)
## Summary & Motivation

Change the semantics of assets_def to only get the assets_def if it was passed in directly, rather than coerced from another format. This is a behavior change. For now, we fall back to the old behavior (in resolve_asset_def) if not found. In 1.11 we will break this.

## How I Tested These Changes

BK

## Changelog

* Changed Definitions.get_assets_def to return an AssetsDefinition without resolving if it was passed in directly. Otherwise it resolves the Definitions object and gets its canonicalze post-resolution representation via resolve_assets_def. The fallback will go away after 1.11