[dg] Rename `--path` context-path-setting option to `--target-path` (#30336)
## Summary 

Motivated by wanting to pass a `--path/-p` argument to
https://linear.app/dagster-labs/issue/BUILD-912/ability-to-filter-the-output-of-dg-list-defs,
this PR changes the `--path` argument which most `dg` commands have to
`--target-path` instead, which is a bit more descriptive of what the
parameter actually does (determines the path for the project/workspace
context to execute, rather than the path to directly "target" by the
command itself).

## Test Plan

Existing tests.