Initial Commit of `ExecutableComponent` (#30403)
## Summary & Motivation

Committing the basic skeleton of the `ExecutableComponent`, which we have used a bit in our internal repo.

We anticipate this becoming a frequently used primitive in the component system. It can eventually serve as a drop-in replacement for `@asset`, `@multi_asset`, `@asset_check`, `@multi_asset_check`. Assuming we land https://github.com/dagster-io/dagster/pull/30070 or something like it `@observable_source_asset` and `@multi_observable_source` asset will also be brought into the fold.

One nice property this has is the clean separation of metadata and compute. 

The user specifies asset specs, asset check specs, and other metadata, and then writes a POPF (Plain 'Ole Python Function) do perform the execution against those specs.

In this way, we can effectively bury the concept of the op, and instead just have vanilla execution functions that bring the asset graph up-to-date.

This will be a WIP for some time as we mature it, but we have found it to be very useful for our own purposes.

This is `ExecutableComponent` as of commit `bbdec964460c8897f38e3a23e508eb527c7d4e53` in `internal`, lightly adapted to fix some imports.

## How I Tested These Changes

BK