Remove partitions_def from ExecutableComponent (#30420)
## Summary & Motivation

In the spirit of only having one way of doing things, I'm going to pull partitions definition of the Executable Component, and rely on it in `AssetSpec` only. Worthy of discussion. If we do this we probably want some group assignment primitive as there are a few properties like this.

## How I Tested These Changes

BK