Make project scoped dg commands stay project scoped, even when they are run in a workspace (#30443)
## Summary & Motivation
Before these commands would run as though you ran them from the workspace root, which is not ideal since some commands can only run in a project context. Change that behavior and allow you to run commands within a single project's context, even if it is part of the workspace.

## How I Tested These Changes
BK, run these commands in a project context and observe changed behavior

## Changelog
[dg] Running dg commands like `dg check defs` and `dg dev` in a project folder that is part of the workspace will now only apply to that project, instead of every project in the workspace.