Remove any extra logging handlers that may have been added at user code import time
Summary:
Workaround for issues like https://github.com/snowplow/snowplow-python-tracker/issues/373 that cause just importing dbt-core to cause unsightly log spew to appear in dg and dagster commands afterwards.

The hesitation here would be if somebody has a legitimate use case for hijacking the root logger? e.g. if they want to emit everything that is logged to a file or something like that? Open to discussion. We could also gate this in some way with an env var or something.

Run dg check defs in a project that imports dbt, no more unsightly logspew

Fixed an issue where importing code that adjusts the handler on the Python root logger would cause extra dagster log commands to be written to the console.