Cron freshness policies in the UI (#30479)
## Summary & Motivation
Support human readable, concise descriptions of cron freshness policies in the UI, in addition to time window freshness policies. Like most of the freshness and asset health UIs, it is Plus-only for now.

Also adds some examples of assets with cron freshness policies to the toys repo.

## How I Tested These Changes
manual UI testing with the toys assets

## Changelog

> Insert changelog entry or delete this section.