[asset health] use loadable by for fetching health data from streamline (#30474)
## Summary & Motivation
Adds `LoadableBy` to the three asset health data classes that are stored
by streamline. In internal, the instance method called will use
`get_multi` to batch fetches into the same query.

Since there are two paths that the resolvers for asset health take
depending on if streamline is enabled, I was considering including the
non-streamline path in the `batch_load` as well, but i'm not 100% sure
that's the right thing to do. This set of changes should be the minimal
work required to get streamline fetches batched together, and i'll
reconsider including the non-streamline path in the `batch_load` in the
future

## How I Tested These Changes
existing unit tests and additional test in
https://github.com/dagster-io/internal/pull/16120, will also observe the
traces in dogfood to confirm that queries are getting batched