Fix default timeout for ExternalRepository call when using dagster dev (#30514)
## Summary & Motivation
The default here should be 180 seconds, but is only set on the real code
server, not the proxy server. So repos that take a long time to run this
call are more likely to time out when using the proxy server.

## How I Tested These Changes
add a time.sleep(75) to ExternalRepository on the server. Before it
would time out, now it does not.

## Changelog
Fixed an issue where `dagster dev` was sometimes failing to load code
locations with a "Deadline Exceeded" error unless the
"--use-legacy-code-server-behavior" flag was set.