Scaffold a local deployment in create-dagster workspace (#30517)
The modified docs in this PR are probably the best explanation for the
changes here. Scaffold a deployments folder alongside projects, with a
local folder that includes a venv that can be used for workspace-level
dg commands. In the future we could add additional deployments for
interacting with deployed dagster instances.