[dg] Include defining module reference as object aliases in the registry (#30551)
## Summary & Motivation

Currently, if I have a component defined at `<root>.components.foo.Foo`,
and the component is imported into `<root>.components` and exposed via
the `<root>.components` entry point, I can't use the defining reference
`<root>.components.foo.Foo`, because this doesn't get entered in the
registry. This PR adds the defining reference as an alias in the
registry, allowing either key to be used.

## How I Tested These Changes

New unit test

## Changelog

defs.yaml files can now specify a component in the module where it is
defined, as opposed to just the module where it is exposed in the `dg`
registry.