[NO_SKIP] Add google-cloud-bigquery pin
Summary:
New version of grpcio triggered some kind of dependency resolution shenigans that forced this down to an older version from 2019 that doesn't work anymore. Add a pin to a version that can actually import with the version of protobuf that dagster uses.

BK (test passes now)

Added a `google-cloud-bigquery>=1.28.3` pin to `dagster-gcp` to correctly reflect the lowest compatible version.

> Insert changelog entry or delete this section.