Add google-cloud-bigquery pin (#30584)
## Summary & Motivation
New version of grpcio triggered some kind of dependency resolution
shenigans that forced this down to an older version from 2019 that
doesn't work anymore. Add a pin to a version that can actually import
with the version of protobuf that dagster uses.

## How I Tested These Changes
BK (test passes now)

## Changelog
Added a `google-cloud-bigquery>=1.28.3` pin to `dagster-gcp` to
correctly reflect the lowest compatible version.