Data version cross-repo fix (#30588)
## Summary & Motivation

Currently, stale status reporting can become incorrect in this scenario:

* Data version system uses
* INITIAL for assets that are assumed to have a value despite no
recorded data version (external assets are always assumed to have a
value)
* NULL for assets that are assumed to have no value (dagster-managed
assets that have never been materialized)
* Suppose we have an asset X and two code locs, A and B. Asset X is
defined as a materializable asset in A and external asset in B.
* Because there is no record for the asset, the framework generates a
default data version.
* In A, this will be NULL (because it is a dagster-managed asset with no
materialization)
* In B, this will be INITIAL (because it is an external asset assumed to
have a value)
* This discrepancy is fine until you introduce another asset Y that
depends on X in code location B. When you materialize Y (in code loc B),
it will save X’s data version as INITIAL in the data provenance
* Then when B’s data version of INITIAL is compared to the NULL version
in the CachingStaleStatusResolver, the discrepancy is interpreted as X
being updated when it never was.

This PR fixes the issue by special-casing the situation when the
previous data version is `INITIAL` and the current data version is
`NULL` in the `CachingStaleStatusResolver`. This is now interpreted as
"not changed".

## How I Tested These Changes

New unit test.

## Changelog

Fixed a bug where assets with cross-location dependencies could
sometimes be incorrectly reported as "Unsynced".