[dg] registry_modules patterns (BUILD-1293) (#30475)
## Summary & Motivation

Add support for wildcards in `project.registry_modules`.

A module pattern looks like this:

- `"*"`
- `"foo.*"`
- `"foo.*.bar"`
- `"foo.bar"` (no `*` is a valid pattern)

The entire module segment has to be "*"-- i.e. this is invalid:
`"foo.*bar.baz"`.

The resolved list of registry modules for a project is generated by
expanding all wildcard patterns.

When a component is scaffolded, its containing module is now only added
to `project.registry_modules` if it doesn't already match a pattern.

Upstack we change the default scaffolded project to include a wildcard
registry module pattern.

## How I Tested These Changes

New unit tests.