[dg] create-dagster update warning (BUILD-1239) (#30573)
## Summary & Motivation

Emit a warning when `create-dagster` is not up to date. This is what it
looks like:


This is adapted from the old code that warned when `dagster-dg` was not
up to date. That code was a little more complicated-- it cached the last
time it checked, and only checked again if an hour had elapsed since the
last check. That's because `dg` gets run all the time.

Since `create-dagster` is intended to be run much less frequently, this
just runs the check every time you do either `create-dagster project` or
`create-dagster workspace`.

## How I Tested These Changes

New unit test

## Changelog

`create-dagster` now warns when scaffolding a project or workspace if it
is not the latest version.