[core] Implement lazy resolution of job configs (#30453)
## Summary & Motivation

This PR fixes [AD-930 - GraphDefinition.to_job() with config blows up if
op has custom IO
manager](https://linear.app/dagster-labs/issue/AD-930/graphdefinitionto-job-with-config-blows-up-if-op-has-custom-io-manager).

Before the changes made in this PR, this code was raising a
`DagsterInvalidDefinitionError` stating that the dummy IO manager was
missing:

```python
from dagster import Definitions, Out, graph, op
from dagster._core.storage.fs_io_manager import FilesystemIOManager


@op(out=Out(io_manager_key="dummy"))
def alpha(x: int):
    return x


@graph
def my_graph(x):
    alpha(x)


my_job = my_graph.to_job(config={"inputs": {"x": 1}})


defs = Definitions(jobs=[my_job], resources={"dummy": FilesystemIOManager(base_dir=".")})
```

The problem: 

When passing a config dictionary to the job, the process that resolves
the configs at job initialization creates a ConfigMapping for the job.
Because the resource definition for the IO manager is passed at the
definition-level and not at the job-level, the resources defs for the
job are incomplete when the configs are resolved.

The solution:

We move the process that resolves the configs from __init__ to the
config properties, so that we resolve configs only the first time they
are used.

## How I Tested These Changes

Updated and new tests with BK.

## Changelog

TODO