[ui] Upgrade to Next 15 (#30536)
## Summary & Motivation

Upgrade app-oss from Next 14 to 15, for https://github.com/dagster-io/dagster/security/dependabot/820.

There doesn't appear to be anything concerning in https://nextjs.org/docs/app/guides/upgrading/version-15 for our use case, and I have found that the upgrade to React 19 is not actually required at this point.

We have a fair amount of work to do to get ourselves onto React 19, given our dependencies, so we shouldn't be blocked on that upgrade.

## How I Tested These Changes

Load the app in JS dev, click around and verify that everything seems to be working normally.

Run a full static build, verify same.