[dg] "plugin"/"library" -> "registry" (#30501)
## Summary & Motivation

Renames a bunch of internal symbols and variables from their "plugin" or
"library" name to a variant with "registry".

Some of these are serdes classes but that's a historical feature
(removed upstack), no longer relevant since we dropped IPC and the
cache. So we can rename them freely.

- `PluginObjectSnap` -> `EnvironmentRegistryObjectSnap`
- `PluginManifest` -> `EnvironmentRegistryManifest`
- `PluginObjectKey` -> `EnvironmentRegistryKey`
- and a few other minor ones

## How I Tested These Changes

Existing test suite.