[dagster-fivetran] Add prefix to op name in build_fivetran_assets_definition (#30598)
## Summary & Motivation

If another op has the exact same name, Dagster will be confused when
parsing the definitions. We experienced this problem in the
dagster-open-platform, where both dlt and Fivetran integration add an op
named "hubspot".

Changes were made in
[internal](https://github.com/dagster-io/internal/pull/16203), but we
should change it here in the asset factory too, to avoid this problem in
the future.

## How I Tested These Changes

Manual test

## Changelog

> Insert changelog entry or delete this section.