Don't fail the build if `make wipe` fails (#30619)
kitchen sink is occassionally failing builds - I think because it can't
do test fixture teardown.

Which is kind of a bummer because we can't use our typical mute tooling
to get around it as easily (other than muting the entire build step).

But I just don't think this cleanup is that important. The dash should
tell make to suppress any errors during this command.

https://www.gnu.org/software/make/manual/html_node/Errors.html


https://buildkite.com/dagster/dagster-dagster/builds/125669#01975abf-306b-4c80-a35f-ebd50f261f59/525-1428