[dg] Remove cache (#30618)
## Summary & Motivation

Code for the `dg` remains in place, but it isn't used anywhere since the
transition to a single-process architecture.

Though it is very possible we will reintroduce some form of caching in
the future, we can safely remove this for now. If we readd the cache
this PR can be looked at for inspiration.

Note that this also removes several user-facing CLI and config file
options.

## How I Tested These Changes

Existing test suite.

## Changelog

The `dg` cache is no longer operative, since `dg` now operates in the
same python environment as the projects it manipulates. Config options
for the cache have been removed.