BUILD-1243: Supporting code for private packages in the integration marketplace (#30527)
## Summary & Motivation

Related: https://github.com/dagster-io/internal/pull/16153

https://linear.app/dagster-labs/issue/BUILD-1243/private-marketplace-demo-technical-poc

This PR lays the foundation for private packages in Dagster Cloud:

- The integration fetching code (getting the list, getting the details
of an integration) is moved to a "provider" hook with a .oss version
- The provider includes stub methods for creating, editing, deleting
packages. When a provider implements these methods, a lightweight editor
UI appears.

- I also renamed "MarketplaceRoot" to "IntegrationsListRoot" and
MarketplaceHome to "IntegrationsListPage" for consistency with other
pages.

## How I Tested These Changes

- Storybooks and manual testing

Co-authored-by: bengotow <bgotow@elementl.com>