Fix issue with "Re-execute from asset failure" (#30027)
## Summary & Motivation

If you do "re-execute from step failure" before "re-execute from asset failure", this does not subset the actual ExecutionPlanSnapshot object, so all the previous steps will still be in there, even though they weren't executed.

This messed with the logic we were using in this function, that just iterated over all steps without checking if they were actually intended to be executed

## How I Tested These Changes

## Changelog

Fixed an issue with the new "re-execute from asset failure" functionality that could cause additional steps to be included if the job was previously re-executed from step failure.