Another tweak to the logic for the backfill daemon when using the threadpool
Summary:
Follow-on to htps://github.com/dagster-io/dagster/commit/c558554129ec4b402c3f13fff520af05e0a70f83. I think that change largely addresses the problem, but even better is to wait to refetch the backfill object until you are certain that the future that is potentially modifying it has finished running.

Launch multiple backfills simultaneously locally with the threadpool on and the thread

> Insert changelog entry or delete this section.