Fix falsey config vals not showing up in launchpad (#30672)
## Summary & Motivation
When you have a "falsey" default value for a piece of config, the
launchpad filters it out. This is because of a very specific exclusion
within the code that was doing a "falsey" check instead of a proper "is
not None" check. Narrowing the condition fixes.

One implication of this is that the actual config field snaps change as
a result. I don't think this is a _breaking_ change, but good to be
aware of.

## How I Tested These Changes
Tested this motivating test case before and after:
```python
from dagster import Config, asset
from pydantic import Field


class NestedConfig(Config):
    field: int | None = Field(default=None)
    field2: int | None = Field(default=None)


class MainConfig(Config):
    nested: NestedConfig = Field(
        default=NestedConfig(
            field=0,
            field2=1000,
        )
    )


@asset(group_name="example")
def example_launchpad_ui(config: MainConfig) -> None:
    print(config.nested.model_dump())

```
Before, field: 0 was not being populated in the launchpad. It is now.

## Note on empty dicts

Empty dicts still do not work here; because there's a block of code that
explicitly filters them out:
https://github.com/dagster-io/dagster/blob/1fa73586c2da28912b84e212dc2f17435b8b9975/python_modules/dagster/dagster/_core/snap/snap_to_yaml.py#L20
I don't know the history here and so am averse to changing it without
@benpankow 's input. My guess is that we need some way to distinguish
between "intentional" empty dicts and "side effect" ones.

## Changelog
- Fixed a bug where "falsey" defualt config values were not showing up
in the launchpad. This has been fixed for all cases except the empty
dictionary.