feat: improve the precision of type annotations for module loaders
This commit uses overloads to improve the type annotations for the
load_assets_from_* functions.

Previously the return type annotation disregarded the effect of the
include_specs function argument. This commit uses overload to improve the type
annotations for the two cases of this boolean being True of False. This commit
also adds the include_specs argument to the function docstrings.