Make custom dbt translators work if you subclass and don't call the constructor
Summary:
Not sure if this constitutes a breaking change or not, but https://github.com/dagster-io/dagster/pull/29123/files added a new property on the translator and not all child classes will neccesarily call the parent constructor.

New test case

Fixed an issue where creating a DagsterDbtTranslator that didn't call the parent class's constructor in it's __init__ method raised an Exception during execution.