[asset health alerting 2/n] refactor AssetHealthStatus enum (#30554)
## Summary & Motivation
This is mainly a refactoring PR.

In a previous PR we introduced `GrapheneAssetHealthStatus` in the
graphene layer. Then later defined the `AssetHealthStatus` python enum.
We can instead derive the `GrapheneAssetHealthStatus` enum from the
python enum rather than define them twice. This also fixes some type
annotations that were failing (couldn't say that
`GrapheneAssetHealthStatus` was the return type of a function, had to
say it was a string).

The graphene layer also held the code that converted the component asset
health statuses into an overall health status. I moved this out of the
GQL layer so that it can be used in other parts of the code base