ASSET_WIPED event (#30049)
## Summary & Motivation
We need to be able to clear the streamline state stored for an asset
when the asset is wiped by a user. This means that we need an event for
streamline to process. This PR does this by emitting an `ASSET_WIPED`
event after the storage method to wipe the event has been called.

Internal companion https://github.com/dagster-io/internal/pull/16081
emits the event in the same transaction as wiping the asset