[perf] [LiveDataProvider] Use hashObject instead of JSON.stringify to lower memory (#30687)
## Summary & Motivation

We're trying to avoid subscribing/re-subscribing to keys if the array
reference changed but the keys are the same. To do that we were using
the keys JSON.stringified as a dependency. Instead, lets use hashObject
to detect if the keys are the same, and use a `useMemo` to change the
keys reference if indeed the hash changed.

## How I Tested These Changes
existing jest tests for LiveDataProvider