[ui] UI Component updates to support billing designs (#30596)
This PR makes a few small changes to the core UI components:

- Our button lists a `fill` prop which we use in billing, but it doesn't
do anything. This PR removes the `fill` prop and aligns the button label
with `text-align: center` so that it renders properly when it's made
wider-than-it's-label via a flexbox.

- Adds an `<Alert intent=“none” />` that renders in gray for a subtle
notice (used for “After your trial ends, you will be billed on Jun 16”
style messages).

- Adds an `<Alert rightButton={<Button…/>} />`. The alerts currently
support displaying an X, but we need to support an actual CTA button
placed inside an alert notice.

- Dialog has a `isCloseButtonShown` prop that is used in several places
but does not actually do anything since we implement our own header.
This is fixed.

## Testing:

The changes above are all exercised in a new billing page and I also
updated stories to demonstrate the new features.

This PR needs to merge with the new billing page because it removes
`<Button fill />`

Co-authored-by: bengotow <bgotow@elementl.com>