[ui] Fix rendering of team tags in asset node tooltips (#30790)
## Summary & Motivation

Related:
https://linear.app/dagster-labs/issue/FE-895/owners-font-in-11019

This PR fixes a rendering issue with the "owners" row in the new faceted
asset nodes. When more than one owner is present, they appear in a
tooltip, and they were not rendered in the the tooltip text color.

I opted to override the tag CSS because the user tags are essentially an
"opaque" component since they are overridden in cloud.

This PR also moves the AssetNode from styled-components to a CSS module.
We talked a bit about these scenarios in the FE sync - in order to
override the rendering of a StyledTag, I applied a global classname to
the component. I think that we'll be doing this for a bunch of
components in order to move off styled-components and this is the first
one.

Testing:
- I also updated the stories to 1) show this facet row by default and 2)
demo both user and team tags, since this bug was just with team tags,
and 3) include stories that test both the single and multiple owner
case.

<img width="365" alt="image"
src="https://github.com/user-attachments/assets/ac3584da-a75a-4ead-9436-14e4b6f97edb"
/>

Co-authored-by: bengotow <bgotow@elementl.com>