[ui] Fix time rendering on asset "recent updates" timeline (#30799)
## Summary & Motivation

When the "recent updates" view on an asset overview falls back to `dagster/last_updated_timestamp` for timestamps, it does not currently take into account that these are second-based timestamps, but the component renders expecting millisecond-based timestamps.

## How I Tested These Changes

View an asset with a recent events view that is known to be broken. Verify that timestamps are correct.

## Changelog

[ui] Fix timestamps on the "Recent events" view on some assets.