[ui] Insights v2: Fix chart clicks (#30828)
## Summary & Motivation

When clicking on an Insights v2 chart, show the dialog based on the x-axis element index being clicked, not the specific y-axis point. This makes it so that when you click on the chart, you don't need to get the click exactly on the graph line.

## How I Tested These Changes

View insights graphs, click to view asset breakdown. Verify that the dialog appears correctly with my needing to click exactly on the graph line.