[DNM] [breaking] Rename FreshnessPolicy to LegacyFreshnessPolicy (#30688)
## Summary & Motivation
Part 1 of moving over to the new `FreshnessPolicy` (currently named `InternalFreshnessPolicy`). We'd like to reuse the `FreshnessPolicy` name, so renaming the existing class to `LegacyFreshnessPolicy`.

Since this is a breaking change for users, it needs to go out in the `1.11` major release. Will need to hold merge to `main` until after the `1.11` release branch is cut and then cherry pick onto the release branch.


## How I Tested These Changes
oss tests, backcompat snapshot test added in the cloud repo: https://github.com/dagster-io/internal/pull/16306

## Changelog

[breaking] rename `FreshnessPolicy` to `LegacyFreshnessPolicy`