[freshness] Use `freshness_policy` parameter to provide `InternalFreshnessPolicy` (#30774)
## Summary & Motivation
https://github.com/dagster-io/dagster/pull/30771 has freed up the `freshness_policy` parameter name, this PR uses that parameter name to pass in `InternalFreshnessPolicy` to `AssetSpec`, asset decorator, etc.

One consequence of this is that the new freshness policy is now stored as a first-class attribute of an asset spec. Prior to this PR, freshness policies were written to asset spec metadata. So we now have to account for both possible locations when reading freshness policies on an asset.

## How I Tested These Changes
existing + new unit tests

## Changelog

> Insert changelog entry or delete this section.
- `freshness_policy` parameter now used to pass the new freshness policies (`InternalFreshnessPolicy`) to asset specs, asset decorator, etc.

(cherry picked from commit 3dfedf40c3b65972fbbb0cf848785aef6a6338d0)