[ui] Check for freshness enabled state in OSS (#30741)
## Summary & Motivation

Move Freshness Policy information for the Asset overview page to be shared components, with evaluation info implementation in Cloud. The OSS pieces of this UI will show a limited amount of information, indicating that these are not evaluated in OSS.

<img width="1406" alt="Screenshot 2025-06-16 at 16 51 14" src="https://github.com/user-attachments/assets/84cead70-890b-4bfa-8c8a-9d95a9ddc881" />

<img width="356" alt="Screenshot 2025-06-16 at 16 51 19" src="https://github.com/user-attachments/assets/e9b01d79-96c9-4543-a42b-7d2c92b92a59" />

## How I Tested These Changes

Run OSS app, navigate to an asset with a freshness policy. Verify that the "Freshness Policy" section no longer renders in the top section of the overview, and that the "Freshness Policy" section in the right panel now shows a "Not evaluated" tag with a tooltip.