[dagster-powerbi-components] Allow content-type translation layer for data sources (#30910)
## Summary

Extends functionality of #29929 to add support for customizing data
source asset keys - these are source assets / deps, so we can't
customize the rest of their properties.

This does raise a broader question of whether we should lock down this
customization on the core translator class (since it presents the user a
false sense of what they can customize)

## How I Tested These Changes

New unit test.