[ui] Add a test spec for useSelectionReducer (#30908)
## Summary & Motivation

I was concerned that updating the list of IDs in `useSelectionReducer` would not result in the selectable IDs being updated correctly, but in fact this does already work because the list and action functions are not memoized.

I added a test spec while working on this, and cleaned up some of the `userEvent` usage and element querying, so I'll just go ahead and land it.

## How I Tested These Changes

TS, lint, jest.