[dg] Make `dg scaffold defs` work for arbitrary valid refs (#30630)
## Summary & Motivation

Currently `dg scaffold defs <ref>` only works for registered objects (or
their aliases, or substring matches). This is part of the legacy of
having a registry sourced from out-of-process.

This PR introduces on the fly dynamic subcommand resolution. If the user
passes:

```
dg scaffold defs some.component.SomeComponent
```

It will work as long as it is a valid python reference to a scaffoldable
object in the environment (though we still rely on the registry when
listing available subcommands).

This will allow removal of the `EnvRegistryObjectSnap` layer upstack.

## How I Tested These Changes

New unit test

---------

Co-authored-by: benpankow <ben@dagsterlabs.com>