Allow assets without definitions to be wiped
Summary:
This fixes an issue where the wipe option was not being shown for external assets in the catalog.

The AssetView.tsx callsite still needs some changes to ensure that the resulting option is actually displayed, since it currently relies on the LaunchAssetExecutionButton component being rendered (and that button is not displayed for these assets).