[1/n] Component decl tree (#30482)
## Summary

Introduces a `ComponentDecl`, which is in effect a closure with all of the information needed to load a component. It typically bundles the `ComponentLoadContext` alongside pointers to the file/file contents which back the component, or its children.

We can build the entire `ComponentDecl` tree without triggering a component load, meaning we can have this tree ready for the load process when e.g. we may invoke a cross-tree load.

Restructures `ComponentTree` to include the structurally similar fns:
- `find_decl_at_path` -> `ComponentDecl`
- `load_component_at_path` -> `Component`
- `build_defs_at_path` -> defs from the component

Each of these is cached. Subsequent PRs will flow all component loads & def building through the `ComponentTree` using these functions, which lets us memoize the tree as it is loaded/defs are built.

## Test Plan

Existing tests + new unit tests.

Update integration tests to validate structure of decl tree.