[2/n] hoist ComponentTree into ComponentLoadContext (#30483)
## Summary

This PR begins the process of recentering component loading on the `ComponentTree` object, which is now "in charge" of the component loading process, including caching components and their definitions. We'll need a back-reference to the tree available in the context so that components can e.g. load other components.

This means any path which constructs a `ComponentLoadContext` will instead construct a `ComponentTree`, which it now uses to vend the context.

This PR is a pretty dumb drop-in replacement for most callsites, but most could subsequently be cleaned up.

## Test Plan

Existing tests.