[3/n] finish consolidating on ComponentTree in load_defs (#30488)
## Summary

Continues the consolidation of manually constructed `ComponentLoadContext`s into `ComponentTree` by handling the `load_defs` case. Currently, this relies on some thorny branching behavior where we try to use `ComponentTree` and fallback to a non-Tree load if the root is not a `DefsFolderComponent`.

This requires some lift so that `ComponentTree` can still be used for e.g. singular components or a single composite YAML file.