convert tuple of groups to a list of groups (#30945)
## Summary & Motivation
I was getting test failure errors like 
```

Full diff:
--
  | [
  | -     AssetSelectionTarget(asset_selection=AndAssetSelection(operands=[GroupsAssetSelection(selected_groups=('the-asset-group',), include_sources=False), CodeLocationAssetSelection(selected_code_location='the-repo@the-location')])),
  | +     AssetSelectionTarget(asset_selection=AndAssetSelection(operands=[GroupsAssetSelection(selected_groups=['the-asset-group'], include_sources=False), CodeLocationAssetSelection(selected_code_location='the-repo@the-location')])),
  | ]

```
important bit 
```
GroupsAssetSelection(selected_groups=('the-asset-group',)
GroupsAssetSelection(selected_groups=['the-asset-group']
```
 
because the group selection gets a list in some situations and a tuple
in some situations. So I switched the tuple to a list. All the tests
seem to work now, so i think it's ok? Not sure if this is actually a big
change to be making

## How I Tested These Changes
existing unit tests