Allow assets without definitions to be wiped (#30941)
Summary:
This fixes an issue where the wipe option was not being shown for external assets in the catalog.

The AssetView.tsx callsite still needs some changes after this to ensure that the resulting dropdown option is actually displayed, since it currently relies on the LaunchAssetExecutionButton component being rendered (and that button is not displayed for these assets).

Changelog:
[dagster-ui] Fixed an issue where a "wipe materializations" option was not being shown in the Dagster UI for assets not defined in code.