[perf][Observe] Persist saved selection subscriptions in memory but no-op them when they're not visible. (#30940)
## Summary & Motivation

Calculating a saved selection and subscribing to health data for it
leads to a lot of expensive operations. On the calculating side we have
to hash the entire asset graph, and on the subscription side we end up
creating individual subscriptions for thousands of assets and mutating
various objects while doing so leading to a lot of CPU usage. Currently
when a selection is rendered (eg. in a tile in catalog home) it has to
do all of this work. The problem is that by switching/collapsing tabs
you can easily get them to render and un-render quickly leading to a lot
of work happening at once and causing a chonky animation. To get around
this we add a provider around the app that will keep instances of the
filtering/health subscriptions alive in memory for the lifetime of the
app. If a selection is not visible then the provider will make
everything no-op using the `skip` parameter of
`useAssetSelectionFiltering` and `useAssetsHealthData` to avoid doing
any expensive work when the selections are not visible. When the
selections are visible again we stop no-oping without needing to setup
any subscriptions or calculate anything from scratch again.


Also stop relying on `useAssetDefinition` because we only need the
definition feeds from the workspace context and that hook returns extra
more expensive fields (eg. metadata)

## How I Tested These Changes




https://github.com/user-attachments/assets/dbff9bbc-4bf0-4231-a87e-b6502916a5f4