[ui] Centralize dayjs extensions and update thresholds FE-881
Users have reported that “a day ago” is too vague and it’d be better to display “23 hours ago” through “47 hours ago” before cutting to two days ago.

This PR centralizes the import of dayjs extensions — even though they’re not all required in all places, they are attached to the global dayjs instance and it’s not valid for components to configure them differently. I think this prevents confusion, and it also allows me to ensure that “47 hours ago” will render consistently everywhere `relativeTIme` was used.